+++ extensions/aligning_guidelines/constant.ts
@@ -6,5 +6,5 @@ export const aligningLineConfig: AligningLineConfig = {
   /** Aligning line dimensions */
   width: 1,
   /** Aligning line color */
-  color: 'rgb(255,0,0,0.9)',
+  color: 'rgba(255,0,0,0.9)',
 };