+++ extensions/aligning_guidelines/typedefs.ts
@@ -1,29 +1,8 @@
-export type VerticalLine = {
-  x: number;
-  y1: number;
-  y2: number;
-};
+import type { Point } from 'fabric';
 
-export type HorizontalLine = {
-  y: number;
-  x1: number;
-  x2: number;
-};
-
-export type VerticalLineProps = {
-  x: number;
-  objectY: number;
-  objectHeight: number;
-  activeObjectY: number;
-  activeObjectHeight: number;
-};
-
-export type HorizontalLineProps = {
-  y: number;
-  objectX: number;
-  objectWidth: number;
-  activeObjectX: number;
-  activeObjectWidth: number;
+export type LineProps = {
+  origin: Point;
+  target: Point;
 };
 
 export type AligningLineConfig = {